+++ R/beta-prime-distribution.R
@@ -41,12 +41,11 @@
 #' @examples 
 #' 
 #' x <- rbetapr(1e5, 5, 3, 2)
-#' xx <- seq(0, 100, by = 0.1)
 #' hist(x, 350, freq = FALSE, xlim = c(0, 100))
-#' lines(xx, dbetapr(xx, 5, 3, 2), col = "red")
+#' curve(dbetapr(x, 5, 3, 2), 0, 100, col = "red", add = TRUE, n = 500)
 #' hist(pbetapr(x, 5, 3, 2))
 #' plot(ecdf(x), xlim = c(0, 100))
-#' lines(xx, pbetapr(xx, 5, 3, 2), col = "red")
+#' curve(pbetapr(x, 5, 3, 2), 0, 100, col = "red", add = TRUE, n = 500)
 #' 
 #' @name BetaPrime
 #' @aliases BetaPrime